added alert dialog before deleting statements

+++ Recorder/app/src/main/java/com/example/recorder/View/StatementAdapter.kt
@@ -1,7 +1,8 @@
 package com.example.recorder.View
 
+import android.app.AlertDialog
+import android.content.DialogInterface
 import android.support.v7.widget.RecyclerView
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import com.example.recorder.Presenter.StatementPresenter
@@ -9,23 +10,34 @@ import com.example.recorder.R
 
 class StatementAdapter : RecyclerView.Adapter<StatementViewHolder> {
 
-    private var presenter : StatementPresenter
+    private var presenter: StatementPresenter
 
-    constructor(presenter: StatementPresenter){
+    constructor(presenter: StatementPresenter) {
         this.presenter = presenter
     }
 
     override fun onCreateViewHolder(p0: ViewGroup, p1: Int): StatementViewHolder {
-        val holder =  StatementViewHolder(
-            LayoutInflater.from(p0.context).inflate(R.layout.statement_list_item,p0,false),
+        val holder = StatementViewHolder(
+            LayoutInflater.from(p0.context).inflate(R.layout.statement_list_item, p0, false),
             presenter
         )
 
-        holder.btnDeleteStatement.setOnClickListener{
-            holder.deleteStatement()
-            Log.i("", "Deleted")
-            notifyDataSetChanged()
-            Log.i("","dataset changed")
+        holder.btnDeleteStatement.setOnClickListener {
+            val builder = AlertDialog.Builder(holder.btnDeleteStatement.context)
+            builder.setMessage(R.string.alertDialogMessage)
+            builder.setCancelable(true)
+            builder.setPositiveButton(
+                R.string.alertDialogButtonPositive
+            ) { _: DialogInterface, _: Int ->
+                holder.deleteStatement()
+                notifyDataSetChanged()
+            }
+            builder.setNegativeButton(
+                R.string.alertDialogButtonNegative
+            ) { _, _ ->
+            }
+            builder.create()
+            builder.show()
         }
         return holder
     }
@@ -35,7 +47,7 @@ class StatementAdapter : RecyclerView.Adapter<StatementViewHolder> {
     }
 
     override fun onBindViewHolder(p0: StatementViewHolder, p1: Int) {
-        presenter.onBindStatementItemView(p0,p1)
+        presenter.onBindStatementItemView(p0, p1)
     }
 
 }
